requirements.txt
streamlit
ollama
easyocr
numpy
Pillow
torch
torchvision
